Pass "ccls.xref" command name as a string, not a symbol, to lsp--send-execute-command in the lsp-execute-command method definition.
This fixes error '(wrong-type-argument json-value-p ccls\.xref)' on attempt
to click on any code lens reference.

It seems that the old elisp-only json serializer did accept a symbol as a JSON
value and serialized it as a string; now, when Emacs is built with native JSON
support, it is not more valid.